Refactor local repository
