+++ app/src/main/java/com/ivy/weatherapp/data/local/WeatherDatabase.kt
@@ -1,12 +1,11 @@
-package com.ivy.weatherapp.data.db
+package com.ivy.weatherapp.data.local
 
 import android.arch.persistence.room.Database
 import android.arch.persistence.room.RoomDatabase
-import com.ivy.weatherapp.data.model.Weather
-import com.ivy.weatherapp.data.repository.WeatherRepository
+import com.ivy.weatherapp.data.local.model.Weather
 
 @Database(entities = [Weather::class], version = 1)
 abstract class WeatherDatabase : RoomDatabase() {
 
-    abstract fun weatherRepository(): WeatherRepository
+    abstract fun weatherRepository(): WeatherDao
 }
