+++ app/src/main/java/com/ivy/weatherapp/data/local/model/Weather.kt
@@ -1,4 +1,4 @@
-package com.ivy.weatherapp.data.model
+package com.ivy.weatherapp.data.local.model
 
 import android.arch.persistence.room.ColumnInfo
 import android.arch.persistence.room.Entity
